+++ file.h
@@ -3,12 +3,12 @@
 
 #include "fileStruct.h"
 
-int runFile(char **input, vari *var);
+int8_t runFile(char **input, vari *var);
 fileStack newFileStack();
-int createTree(char *fileName, fileTree *tree, char **fileString, int *maxSize);
-int executeTree(fileTree *tree, vari *var, int maxSize); 
-int checkProgramFlow(char *input);
-int checkConditional(char *input, int type, vari *var);
+int8_t createTree(char *fileName, fileTree *tree, char **fileString, int *maxSize);
+int8_t executeTree(fileTree *tree, vari *var, int maxSize); 
+int8_t checkProgramFlow(char *input);
+int8_t checkConditional(char *input, int type, vari *var);
 void freeString(char **string, int max);
 
 #endif //FILES
