+++ functions.c
@@ -1,5 +1,6 @@
 #include <math.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #include "operator.h"
 #include "functions.h"
@@ -14,7 +15,7 @@ element factorial(element a) {
 }
 
 
-matrix *eye(matrix *a, int *error){
+matrix *eye(matrix *a, int8_t *error){
   if(a->dimension != 1){
     *error = -12;
     return NULL;
@@ -33,7 +34,7 @@ matrix *eye(matrix *a, int *error){
 }
 
 
-matrix *getSize(matrix *a, int *error){
+matrix *getSize(matrix *a, int8_t *error){
   int newSize[3];
 
   //output is a row vector
@@ -50,11 +51,11 @@ matrix *getSize(matrix *a, int *error){
 }
 
 
-matrix *divideMatrix(matrix *a, matrix *b, int *error){
+matrix *divideMatrix(matrix *a, matrix *b, int8_t *error){
 }
 
 
-matrix *multiplyMatrix(matrix *a, matrix *b, int *error){
+matrix *multiplyMatrix(matrix *a, matrix *b, int8_t *error){
   matrix *out = NULL;
 
   //matrix multiplication only defined for 2d arrays
@@ -94,7 +95,7 @@ matrix *multiplyMatrix(matrix *a, matrix *b, int *error){
 }
 
 
-matrix *exponentMatrix(matrix *a, matrix *b, int *error){
+matrix *exponentMatrix(matrix *a, matrix *b, int8_t *error){
   int aScalar = isScalar(a);
   int bScalar = isScalar(b);
 
@@ -154,7 +155,7 @@ matrix *exponentMatrix(matrix *a, matrix *b, int *error){
 }
 
 
-matrix *transposeMatrix(matrix *a, int *error){
+matrix *transposeMatrix(matrix *a, int8_t *error){
   //transpose only defined for 2d matrix
   if(a->dimension != 2){
     *error = -10;
@@ -184,7 +185,7 @@ matrix *transposeMatrix(matrix *a, int *error){
 
 
 //returns value from one argument functions
-element oneArg(element a, int o, int *error) {
+element oneArg(element a, int o, int8_t *error) {
 switch(o) {
  case eSin: return sin(a);
  case eCos: return cos(a);
@@ -205,7 +206,7 @@ switch(o) {
 
 
 //returns value from two argument function
-element twoArg(element a, element b, int o, int *error) {
+element twoArg(element a, element b, int o, int8_t *error) {
   switch(o) {
   case eAdd: return a + b;
   case eSubtract: return a - b;
