+++ matrix.h
@@ -12,10 +12,10 @@ typedef struct matrix{
 } matrix; 
 
 
-matrix *initMatrix(int *size, int dimension, int *error);
-matrix *initScalar(element e, int *error);
-matrix *copyMatrix(matrix *src, int *error);
-matrix *concatMatrix(matrix *a, matrix *b, int dimension, int *error);
+matrix *initMatrix(int *size, int dimension, int8_t *error);
+matrix *initScalar(element e, int8_t *error);
+matrix *copyMatrix(matrix *src, int8_t *error);
+matrix *concatMatrix(matrix *a, matrix *b, int dimension, int8_t *error);
 
 void freeMatrix(matrix *m);
 void printTwoDMatrix(const matrix m, int offset);
@@ -24,9 +24,9 @@ void printMatrix(const matrix m);
 int getLength(int *size, int dimension);
 int sub2ind(int *location, int *size, int dimension);
 
-int isScalar(matrix *m);
+int8_t isScalar(matrix *m);
 
-int compareSize(int *a, int *b, int dimA, int dimB);
-int checkInnerDim(matrix *a, matrix *b);
+int8_t compareSize(int *a, int *b, int dimA, int dimB);
+int8_t checkInnerDim(matrix *a, matrix *b);
 
 #endif //MATRIX
